refactor: useDeleteAdminApplication을 features/manageApplicant로 분리
