+++ src/features/manageApplicant/index.ts
@@ -1,0 +1,1 @@
+export { useDeleteApplicant } from './model/useDeleteApplicant';
