+++ src/features/manageApplicant/model/useDeleteApplicant.ts
@@ -1,0 +1,17 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { applicationQueryKeys } from '@/entities/application';
+import { applicationUrl, del } from '@/shared/api';
+
+export const useDeleteApplicant = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (id: number) => del<void>(applicationUrl.deleteApplication(id)),
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: applicationQueryKeys.allAdminApplications(),
+      });
+    },
+  });
+};
